LayoutTests/media/modern-media-controls/scrubber-support/scrubber-support-click.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
    video, #host {
        position: absolute;
        top: 0;
        left: 0;
    }

    video {
        width: 800px;
        height: 240px;
    }
    
</style>
<video src="../../content/test.mp4" controls autoplay></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>ScrubberSupport</code> behavior when clicking the scrubber.");

const container = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(container, media, null);

mediaController.controls.showsStartButton = false;

scheduler.frameDidFire = function() {
    if (media.paused)
        return;

    media.pause();

    media.addEventListener("timeupdate", () => {
        shouldBeCloseTo("media.currentTime", 3.1872, 0.2);
        container.remove();
        media.remove();
        finishJSTest();
    });

    const input = mediaController.controls.timeControl.scrubber.children[1].element;
    const bounds = input.getBoundingClientRect();
    const centerX = bounds.left + bounds.width / 2;
    const centerY = bounds.top + bounds.height / 2;

    eventSender.mouseMoveTo(centerX, centerY);
    eventSender.mouseDown();
    eventSender.mouseUp();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
